backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
websockets>=12.0
azure-ai-projects>=2.0.0
azure-identity>=1.17.0
azure-cosmos>=4.7.0
pydantic>=2.0
